Replace atoi with strtol

+++ src/xlsxbook.h
@@ -10,7 +10,7 @@ const std::string& path_;
 std::vector<std::string> sheets_;
 std::vector<std::string> strings_;
 unsigned long int strings_size_;
-std::vector<int> cellXfs_xfId_;
+std::vector<unsigned long int> cellXfs_xfId_;
 
 public:
 
@@ -19,7 +19,7 @@ public:
   std::vector<std::string>& sheets();   // sheet names
   std::vector<std::string>& strings();  // strings table
   unsigned long int strings_size();     // length of strings table
-  std::vector<int>& cellXfs_xfId();     // link from cellXfs from cellStyleXfs
+  std::vector<unsigned long int>& cellXfs_xfId(); // link from cellXfs from cellStyleXfs
 
 private:
